Moved some Subjects to new directory structure
