+++ src/components/UseCommandBar.tsx
@@ -1,13 +1,13 @@
-import { AllRouteComponentProps, Paths } from "../../Routing";
+import { AllRouteComponentProps, Paths } from "../Routing";
 import { CommandBarButton, Toggle } from "office-ui-fabric-react";
 import React, { useCallback, useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
 
-import AppendChildren from "../../views/components/AppendChildren";
-import { BUTTON_HEIGHT } from "../../components/AppCommandBar/Common";
-import SortButton from "../../components/AppCommandBar/SortButton";
-import { State } from "../../Reducer";
-import { setSeparateComplete } from "../../model/Subject/SetSeparateComplete";
+import AppendChildren from "../views/components/AppendChildren";
+import { BUTTON_HEIGHT } from "./AppCommandBar/Common";
+import SortButton from "./AppCommandBar/SortButton";
+import { State } from "../Reducer";
+import { setSeparateComplete } from "../model/Subject/SetSeparateComplete";
 
 const commandBarStyles = { root: { height: BUTTON_HEIGHT } };
