+++ src/subject/components/ListView/UseListViewScroll.ts
@@ -1,7 +1,7 @@
 import { IDetailsList } from "office-ui-fabric-react";
 import { useRef, useEffect } from "react";
 import { getDiffIndex } from "../View";
-import { UseSubjectView } from "../SubjectView";
+import { UseSubjectView } from "../../../components/SubjectView";
 import { SubjectState } from "../../../model/Subject";
 
 interface UseListViewScroll {
